requirement.txt
certifi==2019.11.28
cffi==1.14.0
colorama==0.4.3
cycler==0.10.0
docopt==0.6.2
future==0.18.2
gitdb==4.0.2
GitPython==3.1.0
joblib==0.14.1
jsonpickle==1.3
kiwisolver==1.1.0
matplotlib==3.2.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
munch==2.5.0
numpy==1.18.2
olefile==0.46
packaging==20.3
pandas==1.0.3
Pillow==6.1.0
py-cpuinfo==5.0.0
pycparser==2.20
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.0
six==1.14.0
sklearn==0.0
smmap==3.0.1
torch==1.5.0
torchvision==0.6.0
wrapt==1.12.1